Move check_schema_versions and compose_db_options to athenian.api.db
